QuickLearnerV2/requirements.txt
langchain
langchain-community
langchain-ollama
langchain_huggingface
hf_xet
pypdf
chromadb
pdfminer.six